Develop day16 (#18)
* Add Part 1 for Day 16

Add the code needed for Part 1 of day 16

* Convert to PacketTree class

Create a PacketTree class to hold the decoded packets
Change the main code to use the new class

* Add C# Unit Tests

Add a base sln for C# unit tests
Add unit tests for day 16

* Unit Tests

Change path delimiter to / to fix test failure
Add day 16 project to the base sln file
Rename the test workflow to differentiate between the run tasks

* Remove duplicate files

* Remove sln file
Move tests to use individual projects

* Bump
