+++ 16/dotnet6/code/Day16Input.txt
